task1-recaptcha-stealth/requirements.txt
pytest-playwright
playwright-stealth
